+++ clean_python/base/infrastructure/mapper.py
@@ -1,8 +1,10 @@
+from typing import Any
+
 from ..domain import Json
 
 
 class Mapper:
-    def to_internal(self, external: Json) -> Json:
+    def to_internal(self, external: Any) -> Json:
         return external
 
     def to_external(self, internal: Json) -> Json: